In the loop challenge changed the example: 2**4 becomes 5**3.
During a bootcamp, I was staring at 2**4 and realized it is ambiguous (in the sense that 2**4 == 4**2) and thus can be confusing for students.
It is not a big deal but it is easy to fix so here it is.

NB: I'm not committing the generated md file as it seems to contain more changes (added cstyle attributes) probably due to my ipython version.