Update TDD section in 05-defensive programming
The reason for this change is to encourage users not to write a test function
with many assertions, but rather a separate test function for each assertion.

This makes it easy to introduce testing frameworks and is here illustrated by
calling nosetests when the test functions have been saved to a file.